# docs/swagger.yaml
consumes:
- application/json
definitions:
...
  github_com_etkecc_mrs_internal_model.MatrixError:
    properties:
      errcode:
        description: Matrix error code, e.g M_UNAUTHORIZED
        type: string
      error:
        description: Matrix error message
        type: string
    type: object
...
  github_com_etkecc_mrs_internal_model.QueryDirectoryResponse:
    properties:
      room_id:
        type: string
      servers:
        items:
          type: string
        type: array
    type: object
...
  github_com_etkecc_mrs_internal_model.RoomDirectoryRoom:
    properties:
      avatar_url:
        type: string
      canonical_alias:
        type: string
      guest_can_join:
        type: boolean
      join_rule:
        type: string
      name:
        type: string
      num_joined_members:
        type: integer
      room_id:
        type: string
      room_type:
        type: string
      topic:
        type: string
      world_readable:
        type: boolean
    type: object
  github_com_etkecc_mrs_internal_model.RoomVisibility:
    properties:
      visibility:
        description: always "public"
        type: string
    type: object
...
info:
  contact: {}
  description: Matrix Rooms Search crawls the Matrix federation and serves public-room
    search. It also answers part of the Matrix federation and client-server APIs for
    the rooms it has indexed. Where our implementation diverges from the spec, that
    divergence is documented here on purpose, not hidden. matrixrooms.info is a public
    demo running MRS.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0
  title: Matrix Rooms Search API
  version: 0.0.0
paths:
...
  /_matrix/client/r0/directory/list/room/{room_id}:
    get:
      description: 'Whether a room is listed in the public directory. MRS only indexes
        public rooms, so if we hold the room the answer is "public"; if we have never
        crawled it (or it is banned) you get a 404, same as any homeserver would.
        This is a search engine, not a speakeasy: there is no "private" to report,
        only "we have it" or "we don''t". r0 and v3 share one handler; r0 is the legacy
        alias.'
      parameters:
      - description: Room ID to check
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The room is indexed, and therefore public
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.RoomVisibility'
        "404":
          description: We have never crawled that room
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
      summary: Room visibility in the directory
      tags:
      - matrix-cs
  /_matrix/client/r0/directory/room/{room_alias}:
    get:
      description: 'Turns a human-friendly room alias like #coffee:example.com into
        the actual room ID plus a list of servers that might know where it lives,
        because nobody memorizes a !opaque:id and clients still have to route the
        join somewhere. The r0 and v3 paths are the same handler; r0 is the legacy
        alias we keep alive for clients that never upgraded.'
      parameters:
      - description: 'Room alias to resolve, e.g. #room:server'
        in: path
        name: room_alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Room ID and the servers that can help you join
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.QueryDirectoryResponse'
        "404":
          description: We could not resolve that alias
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
      summary: Resolve a room alias (client)
      tags:
      - matrix-cs
  /_matrix/client/unstable/im.nheko.summary/rooms/{room_id_alias}/summary:
    get:
      description: Room summary (name, topic, member count, join rules) for a room
        ID or alias, per MSC3266. We answer this for rooms we have crawled, so it
        works even though we are a search index and not the room's homeserver. Registered
        at the stable v1 path plus the two unstable im.nheko.summary paths, because
        matrix.to and older clients still call those.
      parameters:
      - description: Room ID or alias to summarize
        in: path
        name: room_id_alias
        required: true
        type: string
      - description: A server to try for the summary if we do not hold the room locally.
          Only the first value is used.
        in: query
        name: via
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Room summary
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.RoomDirectoryRoom'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
      summary: Room summary
      tags:
      - matrix-cs
  /_matrix/client/unstable/im.nheko.summary/summary/{room_id_alias}:
    get:
      description: Room summary (name, topic, member count, join rules) for a room
        ID or alias, per MSC3266. We answer this for rooms we have crawled, so it
        works even though we are a search index and not the room's homeserver. Registered
        at the stable v1 path plus the two unstable im.nheko.summary paths, because
        matrix.to and older clients still call those.
      parameters:
      - description: Room ID or alias to summarize
        in: path
        name: room_id_alias
        required: true
        type: string
      - description: A server to try for the summary if we do not hold the room locally.
          Only the first value is used.
        in: query
        name: via
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Room summary
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.RoomDirectoryRoom'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
      summary: Room summary
      tags:
      - matrix-cs
  /_matrix/client/v1/room_summary/{room_id_alias}:
    get:
      description: Room summary (name, topic, member count, join rules) for a room
        ID or alias, per MSC3266. We answer this for rooms we have crawled, so it
        works even though we are a search index and not the room's homeserver. Registered
        at the stable v1 path plus the two unstable im.nheko.summary paths, because
        matrix.to and older clients still call those.
      parameters:
      - description: Room ID or alias to summarize
        in: path
        name: room_id_alias
        required: true
        type: string
      - description: A server to try for the summary if we do not hold the room locally.
          Only the first value is used.
        in: query
        name: via
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Room summary
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.RoomDirectoryRoom'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
      summary: Room summary
      tags:
      - matrix-cs
  /_matrix/client/v3/directory/list/room/{room_id}:
    get:
      description: 'Whether a room is listed in the public directory. MRS only indexes
        public rooms, so if we hold the room the answer is "public"; if we have never
        crawled it (or it is banned) you get a 404, same as any homeserver would.
        This is a search engine, not a speakeasy: there is no "private" to report,
        only "we have it" or "we don''t". r0 and v3 share one handler; r0 is the legacy
        alias.'
      parameters:
      - description: Room ID to check
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The room is indexed, and therefore public
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.RoomVisibility'
        "404":
          description: We have never crawled that room
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
      summary: Room visibility in the directory
      tags:
      - matrix-cs
  /_matrix/client/v3/directory/room/{room_alias}:
    get:
      description: 'Turns a human-friendly room alias like #coffee:example.com into
        the actual room ID plus a list of servers that might know where it lives,
        because nobody memorizes a !opaque:id and clients still have to route the
        join somewhere. The r0 and v3 paths are the same handler; r0 is the legacy
        alias we keep alive for clients that never upgraded.'
      parameters:
      - description: 'Room alias to resolve, e.g. #room:server'
        in: path
        name: room_alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Room ID and the servers that can help you join
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.QueryDirectoryResponse'
        "404":
          description: We could not resolve that alias
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
      summary: Resolve a room alias (client)
      tags:
      - matrix-cs
...
  /_matrix/media/r0/thumbnail/{name}/{id}:
    get:
      description: Serves a media thumbnail. Tries the unauthenticated client-server
        media API first (faster), then falls back to the authenticated federation
        media API. Streams the image on a hit, or 204 when there is nothing to serve.
        Standard Matrix thumbnail query params (width, height, method) are passed
        straight through.
      parameters:
      - description: Media origin server name
        in: path
        name: name
        required: true
        type: string
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      - description: Thumbnail width
        in: query
        name: width
        type: integer
      - description: Thumbnail height
        in: query
        name: height
        type: integer
      - description: 'Resize method: crop or scale'
        in: query
        name: method
        type: string
      produces:
      - image/jpeg
      - image/png
      responses:
        "200":
          description: Thumbnail image
          schema:
            type: file
        "204":
          description: No thumbnail available
      summary: Media thumbnail
      tags:
      - media
  /_matrix/media/v3/thumbnail/{name}/{id}:
    get:
      description: Serves a media thumbnail. Tries the unauthenticated client-server
        media API first (faster), then falls back to the authenticated federation
        media API. Streams the image on a hit, or 204 when there is nothing to serve.
        Standard Matrix thumbnail query params (width, height, method) are passed
        straight through.
      parameters:
      - description: Media origin server name
        in: path
        name: name
        required: true
        type: string
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      - description: Thumbnail width
        in: query
        name: width
        type: integer
      - description: Thumbnail height
        in: query
        name: height
        type: integer
      - description: 'Resize method: crop or scale'
        in: query
        name: method
        type: string
      produces:
      - image/jpeg
      - image/png
      responses:
        "200":
          description: Thumbnail image
          schema:
            type: file
        "204":
          description: No thumbnail available
      summary: Media thumbnail
      tags:
      - media
...
produces:
- application/json
schemes:
- https
securityDefinitions:
  AdminAuth:
    type: basic
  CatalogAuth:
    type: basic
  DiscoveryAuth:
    type: basic
  FederationAuth:
    description: 'Matrix federation signature auth: an `Authorization: X-Matrix origin=...,key=...,sig=...`
      header carrying an ed25519 signature over the request, which we verify against
      the origin server''s published keys.'
    in: header
    name: Authorization
    type: apiKey
  MetricsAuth:
    type: basic
  ModerationAuth:
    type: basic
swagger: "2.0"